Moved ExchangeRate.cs to folder utils.
