+++ RebelAllianceBank/Enums/EnumsExchangeRate.cs
